requirements.txt
fastapi==0.115.5
requests==2.32.3
python-dotenv==1.0.1
uvicorn==0.32.1
python-multipart==0.0.18
playwright==1.49.0
beautifulsoup4
tqdm

## is neccesary run playwright install:
# playwright install